+++ payment-api/src/main/java/lol/maki/socks/config/PaymentProps.java
@@ -1,18 +1,20 @@
 package lol.maki.socks.config;
 
+import java.math.BigDecimal;
+
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.ConstructorBinding;
 
 @ConfigurationProperties(prefix = "payment")
 @ConstructorBinding
 public class PaymentProps {
-	private final float declineOverAmount;
+	private final BigDecimal declineOverAmount;
 
-	public PaymentProps(float declineOverAmount) {
+	public PaymentProps(BigDecimal declineOverAmount) {
 		this.declineOverAmount = declineOverAmount;
 	}
 
-	public float declineOverAmount() {
+	public BigDecimal declineOverAmount() {
 		return declineOverAmount;
 	}
 }
